Identifier unit tests
